+++ Internal/MedProcessTools/MedProcessTools/FeatureProcess.cpp
@@ -598,7 +598,7 @@ int FeatureImputer::Learn(MedFeatures& features, unordered_set<int>& ids) {
 		if (strata_sizes[i] < min_samples) { // Not enough values to make valid imputation
 			too_small_stratas++;
 			if (moment_type == IMPUTE_MMNT_SAMPLE)
-				histograms[i].push_back({ missing_value,1.0 });
+				histograms[i].push_back({ missing_value,(float)1.0 });
 			else
 				moments[i] = missing_value;
 		}
